Converted TeacherWebSocketService to Angular. #2466
