+++ src/main/webapp/wise5/classroomMonitor/classroomMonitorComponents/nodeProgress/navItem/navItem.ts
@@ -5,7 +5,7 @@ import { ConfigService } from '../../../../services/configService';
 import NotificationService from '../../../../services/notificationService';
 import { StudentStatusService } from '../../../../services/studentStatusService';
 import TeacherDataService from '../../../../services/teacherDataService';
-import TeacherWebSocketService from '../../../../services/teacherWebSocketService';
+import { TeacherWebSocketService } from '../../../../services/teacherWebSocketService';
 import * as $ from 'jquery';
 import { TeacherProjectService } from '../../../../services/teacherProjectService';
