+++ src/main/webapp/wise5/classroomMonitor/studentProgress/studentProgressController.ts
@@ -3,7 +3,7 @@
 import { ConfigService } from '../../services/configService';
 import { StudentStatusService } from '../../services/studentStatusService';
 import TeacherDataService from '../../services/teacherDataService';
-import TeacherWebSocketService from '../../services/teacherWebSocketService';
+import { TeacherWebSocketService } from '../../services/teacherWebSocketService';
 
 class StudentProgressController {
   currentWorkgroup: any;
